+++ backend/src/main/java/com/carffeine/carffeine/station/controller/auth/dto/MemberResponse.java
@@ -1,0 +1,4 @@
+package com.carffeine.carffeine.station.controller.auth.dto;
+
+public record MemberResponse(Long id, String email) {
+}
