--- backend/src/main/java/com/carffeine/carffeine/station/controller/security/dto/MemberResponse.java
@@ -1,4 +1,0 @@
-package com.carffeine.carffeine.station.controller.security.dto;
-
-public record MemberResponse(Long id, String email) {
-}
