+++ backend/src/test/java/com/carffeine/carffeine/station/MockBeanInjection.java
@@ -3,8 +3,8 @@ package com.carffeine.carffeine.station;
 import com.carffeine.carffeine.station.domain.jwt.JwtProvider;
 import com.carffeine.carffeine.station.domain.user.UserRepository;
 import com.carffeine.carffeine.station.infrastructure.api.AuthMemberResolver;
+import com.carffeine.carffeine.station.service.auth.UserService;
 import com.carffeine.carffeine.station.service.congestion.CongestionService;
-import com.carffeine.carffeine.station.service.security.UserService;
 import com.carffeine.carffeine.station.service.station.StationService;
 import org.springframework.boot.test.mock.mockito.MockBean;
